add special features support

+++ main.py
@@ -11,6 +11,8 @@ import threading
 from pathlib import Path
 from tkinter import simpledialog
 from dotenv import load_dotenv
+from os import listdir
+from os.path import isfile, join
 
 load_dotenv()
 drive_number = 0
@@ -20,6 +22,11 @@ TMDB_API_KEY = os.getenv("TMDB_API_KEY")
 release_year = None
 movie_name = None
 movie_id = None
+listbox_extra = None
+output_folder_extra = None
+extra_disc = None
+movie = None
+root_extra = None
 discord = True
 DISCORD_WEBHOOK = os.getenv("DISCORD_WEBHOOK")
 DISCORD_USER_ID = os.getenv("DISCORD_USER_ID")
@@ -49,33 +56,68 @@ def get_disc_label(device):
     return None
 
 def select_movie(event=None):
-    global release_year, movie_name, movie_id
+    global release_year, movie_name, movie_id, listbox_extra, output_folder_extra, extra_disc, root_extra
     selection = listbox.curselection()
     if selection:
-        index = selection[0]
-        movie = movies_data[index]
+        if selection == (0,):
+            extra_disc = True
+            root.destroy()
+            root_extra = tk.Tk()
+            root_extra.title("Selecteer de film?")
+            listbox_extra = tk.Listbox(root_extra, height=15, width=60)
+            listbox_extra.pack(pady=10)
+            output_folder_extra = [f for f in listdir(output_folder_root)]
+            for movie in output_folder_extra:
+                listbox_extra.insert(tk.END, movie)
+            listbox_extra.bind("<Double-1>", extra_selecteren)
+            listbox_extra.bind("<Return>", extra_selecteren)
+        else:
+            extra_disc = False
+            index = selection[0]
+            index = index - 1
+            movie = movies_data[index]
 
-        # los opslaan
-        release_year = movie["release_date"].split("-")[0]
-        movie_name = movie["title"]
-        movie_id = movie["id"]
+            # los opslaan
+            release_year = movie["release_date"].split("-")[0]
+            movie_name = movie["title"]
+            movie_id = movie["id"]
 
-        root.destroy()  # sluit venster
+            root.destroy()  # sluit venster
+
+def extra_selecteren(event=None):
+    global movie, root_extra
+    selection = listbox_extra.curselection()
+    index = selection[0]
+    movie = output_folder_extra[index]
+    root_extra.destroy()
+
 
 def handbrake(folder):
-    os.rename(folder, f"{folder}_old")
-    os.makedirs(folder)
-    os.makedirs(f"{folder}/extras")
-    directory = Path(f"{folder}_old")
-    for file in directory.rglob("*"):  # rglob allows recursion
-        if file.is_file():
-            print(f"Transcoding: {file}")
-            if "/extras/" in str(file).replace("\\", "/"):
-                extras_file = Path(file)
-                subprocess.run(["flatpak", "run", "--command=HandBrakeCLI", "fr.handbrake.ghb", "-i", file, "-o", f"{folder}/extras/{extras_file.name}", "--preset-import-file", "preset.json"])
-            else:
+    print(f"Starting Handbrake with extras: {extra_disc}")
+    if extra_disc:
+        print(f"Starting Handbrake with extras and {folder}")
+        directory = Path(folder)
+        for file in directory.rglob("*"):  # rglob allows recursion
+            if file.is_file():
+                print(f"Transcoding: {file}")
                 file_path = Path(file)
-                subprocess.run(["flatpak", "run", "--command=HandBrakeCLI", "fr.handbrake.ghb", "-i", file, "-o", f"{folder}/{file_path.name}", "--preset-import-file", "preset.json"])
+                subprocess.run(["flatpak", "run", "--command=HandBrakeCLI", "fr.handbrake.ghb", "-i", file, "-o", f"{output_folder_root}/{movie}/extras/{file_path.name}", "--preset-import-file", "preset.json"])
+        shutil.remove(folder)
+    else:
+        os.rename(folder, f"{folder}_old")
+        os.makedirs(folder)
+        os.makedirs(f"{folder}/extras")
+        directory = Path(f"{folder}_old")
+        for file in directory.rglob("*"):  # rglob allows recursion
+            if file.is_file():
+                print(f"Transcoding: {file}")
+                if "/extras/" in str(file).replace("\\", "/"):
+                    extras_file = Path(file)
+                    subprocess.run(["flatpak", "run", "--command=HandBrakeCLI", "fr.handbrake.ghb", "-i", file, "-o", f"{folder}/extras/{extras_file.name}", "--preset-import-file", "preset.json"])
+                    
+                else:
+                    file_path = Path(file)
+                    subprocess.run(["flatpak", "run", "--command=HandBrakeCLI", "fr.handbrake.ghb", "-i", file, "-o", f"{folder}/{file_path.name}", "--preset-import-file", "preset.json"])
 
 while True:
     print("Waiting for Disc")
@@ -117,6 +159,7 @@ while True:
         root.title("Selecteer de film?")
         listbox = tk.Listbox(root, height=15, width=60)
         listbox.pack(pady=10)
+        listbox.insert(tk.END, "Extra")
         for movie in movies_data:
             listbox.insert(tk.END, f"{movie["title"]} ({movie["release_date"].split("-")[0]})")
         listbox.bind("<Double-1>", select_movie)
@@ -132,6 +175,7 @@ while True:
         root.title("Selecteer de film?")
         listbox = tk.Listbox(root, height=15, width=60)
         listbox.pack(pady=10)
+        listbox.insert(tk.END, "Extra")
         for movie in movies_data:
             listbox.insert(tk.END, f"{movie["title"]} ({movie["release_date"].split("-")[0]})")
         listbox.bind("<Double-1>", select_movie)
@@ -154,6 +198,7 @@ while True:
             root.title("Selecteer de film?")
             listbox = tk.Listbox(root, height=15, width=60)
             listbox.pack(pady=10)
+            listbox.insert(tk.END, "Extra")
             for movie in movies_data:
                 listbox.insert(tk.END, f"{movie["title"]} ({movie["release_date"].split("-")[0]})")
             listbox.bind("<Double-1>", select_movie)
@@ -167,40 +212,46 @@ while True:
             print("Release jaar:", release_year)
             print("ID:", movie_id)
 
-    output_folder = f"{movie_name} ({release_year}) [tmdbid-{movie_id}]"
+    if extra_disc:
+        output_folder = "extras_disc"
+    else:
+        output_folder = f"{movie_name} ({release_year}) [tmdbid-{movie_id}]"
 
     if not os.path.exists(f"{output_folder_root}/temp_folder"):
         os.makedirs(f"{output_folder_root}/temp_folder")
 
 
-    subprocess.run(["makemkvcon", f"--minlength={minlength_title}", "mkv", f"disc:{drive_number}", "all", f"{output_folder_root}/temp_folder"])
+    #subprocess.run(["makemkvcon", f"--minlength={minlength_title}", "mkv", f"disc:{drive_number}", "all", f"{output_folder_root}/temp_folder"])
     item_path = f"{output_folder_root}/temp_folder"
     new_dir_name = output_folder
     new_dir_path = os.path.join(output_folder_root, new_dir_name)
     extras_dir = os.path.join(new_dir_path, 'extras')
+
+    if not extra_disc:
+        os.makedirs(new_dir_path, exist_ok=True)
+        os.makedirs(extras_dir, exist_ok=True)
+        files = [f for f in os.listdir(item_path) if os.path.isfile(os.path.join(item_path, f))]
+        if not files:
+            continue
+        
+        largest_file = max(files, key=lambda x: os.path.getsize(os.path.join(item_path, x)))
+        
+        ext = os.path.splitext(largest_file)[1]
+        new_video_name = f"{new_dir_name}{ext}"
+        
+        try:
+            shutil.move(os.path.join(item_path, largest_file), os.path.join(new_dir_path, new_video_name))
+
+            if not extra_disc:
+                for f in files:
+                    if f == largest_file:
+                        continue
+                    shutil.move(os.path.join(item_path, f), os.path.join(extras_dir, f))
             
-    os.makedirs(new_dir_path, exist_ok=True)
-    os.makedirs(extras_dir, exist_ok=True)
-    files = [f for f in os.listdir(item_path) if os.path.isfile(os.path.join(item_path, f))]
-    if not files:
-        continue
-    
-    largest_file = max(files, key=lambda x: os.path.getsize(os.path.join(item_path, x)))
-    
-    ext = os.path.splitext(largest_file)[1]
-    new_video_name = f"{new_dir_name}{ext}"
-    
-    try:
-        shutil.move(os.path.join(item_path, largest_file), os.path.join(new_dir_path, new_video_name))
-        
-        for f in files:
-            if f == largest_file:
-                continue
-            shutil.move(os.path.join(item_path, f), os.path.join(extras_dir, f))
-        
-        os.rmdir(item_path)
-    except Exception as e:
-        messagebox.showerror("Error", f"Failed to process {item}: {str(e)}")
+            os.rmdir(item_path)
+        except Exception as e:
+            messagebox.showerror("Error", f"Failed to process {item}: {str(e)}")
+
     subprocess.run(["eject", f"/dev/sr{drive_number}"])
     if discord:
         webhook_url = DISCORD_WEBHOOK
@@ -213,7 +264,10 @@ while True:
         # Send the POST request
         response = requests.post(webhook_url, json=data)
     if handbrake_flatpak:
-        handbrake_folder = f"{output_folder_root}/{output_folder}"
+        if extra_disc:
+            handbrake_folder = f"{output_folder_root}/temp_folder"
+        else:
+            handbrake_folder = f"{output_folder_root}/{output_folder}"
         threading.Thread(
             target=handbrake,
             args=(handbrake_folder,),
